requirements.txt
absl-py==1.0.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anykeystore==0.2
apex==0.9.10.dev0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
black==21.4b2
blis==0.7.11
boto3==1.28.62
botocore==1.31.62
cachetools==4.2.4
catalogue==2.0.10
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cloudpathlib==0.15.1
cloudpickle==2.0.0
cmake==3.27.6
confection==0.1.3
cryptacular==1.6.2
cycler==0.11.0
cymem==2.0.8
Cython==0.29.24
deepspeed==0.11.0
defusedxml==0.7.1
detectron2 @ file:///root/detectron2/dist/detectron2-0.6-cp38-cp38-linux_x86_64.whl
docker-pycreds==0.4.0
exceptiongroup==1.1.3
expecttest==0.1.3
FCD==1.2
filelock==3.12.4
fonttools==4.28.2
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.2
fvcore==0.1.5.post20211023
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.2
google-auth-oauthlib==1.0.0
graphviz==0.20.1
greenlet==3.0.0
grpcio==1.59.0
hjson==3.1.0
huggingface-hub==0.17.3
hupper==1.12
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==2.0.0
iopath==0.1.9
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.3.2
langcodes==3.3.0
Levenshtein==0.23.0
lightning-utilities==0.9.0
lit==17.0.2
littleutils==0.2.2
lmdb==1.2.1
Markdown==3.3.6
MarkupSafe==2.1.3
matplotlib==3.5.0
megatron-lm @ file:///tmp/MolBART/megatron_molbart/Megatron-LM-v1.1.5-3D_parallelism
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==0.4.3
networkx==3.1
ninja==1.11.1
nltk==3.8.1
numpy==1.21.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.1
ogb==1.3.6
omegaconf==2.1.1
outdated==0.2.2
packaging==21.3
pandas==2.0.3
PasteDeploy==3.0.1
pathspec==0.9.0
pathtools==0.1.2
pathy==0.10.2
pbkdf2==1.3
Pillow==8.4.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
pluggy==1.3.0
portalocker==2.3.2
preshed==3.0.9
protobuf==4.24.4
psutil==5.8.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.11.1
pycocotools==2.0.3
pydantic==1.10.13
pydantic_core==2.10.1
pyDeprecate==0.3.2
pydot==1.4.2
pyparsing==3.0.6
pyramid==2.0.2
pyramid-mailer==0.15.1
pytest==7.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytorch-lightning==1.7.0
pytz==2023.3.post1
PyYAML==6.0
rapidfuzz==3.4.0
rdkit==2023.3.3
regex==2021.11.10
repoze.sendmail==4.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rouge-score==0.1.2
rsa==4.8
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
s3transfer==0.7.0
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.10.1
sentencepiece==0.1.99
sentry-sdk==1.31.0
setproctitle==1.3.3
setuptools-scm==6.3.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.21
srsly==2.4.8
sympy==1.12
tabulate==0.8.9
tencent-light==3.1.1
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
thinc==8.2.1
threadpoolctl==3.2.0
tokenizers==0.14.1
toml==0.10.2
tomli==1.2.2
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
torchviz==0.0.2
tqdm==4.62.3
transaction==3.1.0
transformers==4.34.0
translationstring==1.4
triton==2.0.0
typer==0.9.0
types-dataclasses==0.6.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.17
velruse==1.1.1
venusian==3.0.0
wandb==0.15.12
wasabi==1.1.2
weasel==0.3.2
WebOb==1.8.7
Werkzeug==2.0.2
WTForms==3.0.1
wtforms-recaptcha==0.3.2
xformers==0.0.22
yacs==0.1.8
yarl==1.9.2
zipp==3.6.0
zope.deprecation==5.0
zope.interface==6.1
zope.sqlalchemy==3.1
